use FakeRpcCommunicator in Finder test
